Create the base of a metrics component

+++ src/Tolerance/MessageProfile/Storage/Neo4jStorage.php
@@ -154,6 +154,6 @@ class Neo4jStorage implements ProfileStorage
      */
     private function getUnixTimestamp(\DateTimeInterface $date)
     {
-        return (double) $date->format('U.u');
+        return (float) $date->format('U.u');
     }
 }
